# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IntelliDesk API
  version: 0.1.0
  description: |
    与 PRD_Backend_Detailed 对齐。鉴权：请求头 `X-Tenant-Id`（必填）、`X-User-Id`（可选，演示用已存在用户 id）。
servers:
  - url: http://localhost:4000
    description: 本地后端直连（默认 PORT）
  - url: http://localhost:4001
    description: 经前端 Vite 同源代理（npm run dev:api，浏览器站点根）
...
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - TenantHeader: []
paths:
...
  /api/after-sales:
    get:
      tags: [after-sales]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: list }
    post:
      tags: [after-sales]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '201': { description: created }
  /api/after-sales/{id}:
    get:
      tags: [after-sales]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
    patch:
      tags: [after-sales]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
...
  /api/dashboard/structure:
    get:
      tags: [dashboard]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: 结构分布 }
  /api/dashboard/trends:
    get:
      tags: [dashboard]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: range
          in: query
          schema: { type: string, enum: [7d, 30d] }
      responses:
        '200': { description: 趋势序列 }
...